% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTD_bootstrap.glm.R
\name{PTD_bootstrap.glm}
\alias{PTD_bootstrap.glm}
\title{Predict-Then-Debias bootstrap for generalized linear models}
\usage{
PTD_bootstrap.glm(
  true_data_completeSamp,
  predicted_data_completeSamp,
  predicted_data_incompleteSamp,
  regFormula.glm,
  GLM_type,
  B = 2000,
  alpha = 0.05,
  TuningScheme = "Diagonal",
  clustered = FALSE,
  stratified = FALSE,
  speedup = FALSE,
  prob_lab_completeSamp = NULL,
  prob_lab_incompleteSamp = NULL,
  GroupID_completeSamp = NULL,
  GroupID_incompleteSamp = NULL,
  TotalStrataSizes = NULL
)
}
\arguments{
\item{true_data_completeSamp}{An n x p data frame with the gold standard data.}

\item{predicted_data_completeSamp}{An n x p data frame where all variables that are not widely available are replaced by their widely available proxies}

\item{predicted_data_incompleteSamp}{An (N-n) x p data where all variables that are not widely available are replaced by their widely available proxies. This dataset corresponds to the samples where gold standard measurement of some variables are unavailable.}

\item{regFormula.glm}{A character string giving the regression specification. Standard formatting that is used for the \code{lm()} and \code{glm()} functions in R should be used.}

\item{GLM_type}{A character string that should be set to either "linear", "logistic", or "Poisson" to indicate whether the estimands of interest are the regression coefficients of a linear regression model, logistic regression model or a Poisson regression model, respectively.}

\item{B}{The number of bootstrap draws B (the default is 2000. Larger B cannot hurt, but can lead to slower runtime).}

\item{alpha}{Statistical significance level. The default is 0.05, but note that the corresponding paper, Kluger et al. (2025), used 0.1.}

\item{TuningScheme}{Character string of tuning schemes to use. "Optimal" uses an estimate of the optimal tuning matrix."Diagonal" (the default) uses an estimate of the optimal tuning matrix among diagonal tuning matrices. "None" sets the tuning matrix to the identity matrix and corresponds to the untuned Predict-Then-Debias estimator.}

\item{clustered}{A TRUE/FALSE logical of whether to use the clustered Predict-Then-Debias Bootstrap (Algorithm 5 in  Kluger et al. (2025)) to account for data that was labelled according to cluster sampling.}

\item{stratified}{A TRUE/FALSE logical of whether to use the stratified Predict-Then-Debias Bootstrap (Algorithm 6 in  Kluger et al. (2025)) to account for data that was collected according to stratified sampling (with a small number of large strata).}

\item{speedup}{A TRUE/FALSE logical of whether to speedup the bootstrap procedure using Algorithm 3 in Kluger et al. (2025).}

\item{prob_lab_completeSamp}{An n x 1 vector with the probabilities that each complete sample would have been labelled and therefore assigned to the complete sample  (default is NULL).}

\item{prob_lab_incompleteSamp}{An (N-n) x 1 vector with the probabilities that each incomplete sample would have been labelled and therefore assigned to the complete sample  (default is NULL).}

\item{GroupID_completeSamp}{An n x 1 vector of integers or character strings giving the corresponding Cluster IDs (if clustered=TRUE) or the Strata IDs (if stratified=TRUE) for each of the n complete samples. When clustered=FALSE and stratified=FALSE, this variable is ignored.}

\item{GroupID_incompleteSamp}{An (N-n) x 1 vector of integers or character strings giving the corresponding Cluster IDs (if clustered=TRUE) or the Strata IDs (if stratified=TRUE) for each of the N-n incomplete samples. When clustered=FALSE and stratified=FALSE, this variable is ignored.}

\item{TotalStrataSizes}{A named vector with the number of elements in each strata from the large population from which the stratified sample was obtained. The names should correspond to the unique strata IDs and the values should correspond to the overall strata sizes. Only need to set this when stratified=TRUE.}
}
\value{
Returns a list with

-The Predict-Then-Debias estimate of the regression coefficients.

-The (1-alpha) x 100 percent confidence intervals for the regression coefficients (based on the Predict-Then-Debias bootstrap).

-The tuning matrix that was ultimately used (this will typically be an estimate of the optimal or optimal diagonal tuning matrix).
}
\description{
Implementation of the Predict-Then-Debias bootstrap for settings where
the estimand is the regression coefficient in a generalized linear model (GLM).
This function calls more general functions such as \code{PTD_bootstrap()}
and does not require the user to specify a function as an input argument. PLEASE NOTE: the first 3 arguments, predicted_data_completeSamp, predicted_data_incompleteSamp, and true_data_completeSamp should all be dataframes with the exact same column names.
}
